requirements.txt
asgiref==3.8.1
boto3==1.38.14
botocore==1.38.14
Django==5.2
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.0
et_xmlfile==2.0.0
jmespath==1.0.1
Markdown==3.8
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
